// src/app/shared/swagger-generated/model/userDTO.ts
/**
 * TicTacToe API
 * TicTacToe API documentation
 *
 * OpenAPI spec version: 1.0.0
 * Contact: hlavjah@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface UserDTO { 
    activated?: boolean;
    authorities?: Array<string>;
    createdBy?: string;
    createdDate?: Date;
    email?: string;
    firstName?: string;
    id?: number;
    imageUrl?: string;
    langKey?: string;
    lastModifiedBy?: string;
    lastModifiedDate?: Date;
    lastName?: string;
    login?: string;
}